+++ client/package.json
@@ -24,6 +24,7 @@
     "@angular/platform-browser": "~8.2.4",
     "@angular/platform-browser-dynamic": "~8.2.4",
     "@angular/router": "~8.2.4",
+    "@auth0/angular-jwt": "^3.0.0",
     "@ngrx/store": "^7.4.0",
     "ng-swagger-gen": "^1.8.1",
     "rxjs": "~6.4.0",
