--- client/src/app/app.component.scss
@@ -1,11 +1,0 @@
-.topAppBar {
-    z-index: 7;
-}
-
-.main-content {
-    width: 90%;
-}
-
-.drawer {
-    position: fixed;
-}
