+++ client/src/app/components/layout/main/home.component.html
@@ -1,0 +1,15 @@
+<div #main>
+    <mdc-top-app-bar class="topAppBar" fixed [scrollTarget]="content">
+        <app-top-bar></app-top-bar>
+    <mdc-linear-progress [open]="isBusy"></mdc-linear-progress>
+    </mdc-top-app-bar>
+    <mdc-drawer #drawer class="drawer" mdcTopAppBarFixedAdjust [drawer]="main.offsetWidth< 1200? 'modal' : 'dismissible'" [open]="isOpen" (openedChange)="onSidebarChanged($event)" (drawerChange)="onDrawerChange(drawer.drawer)">
+        <mdc-drawer-header title="Tim's Homeserver" subtitle="Apps"></mdc-drawer-header>
+        <app-side-bar></app-side-bar>
+    </mdc-drawer>
+    <div mdcTopAppBarFixedAdjust mdcDrawerAppContent #content fxLayout="row" fxLayoutAlign="center">
+        <div fxLayout="column" class="main-content">
+            <router-outlet></router-outlet>
+        </div>
+    </div>
+</div>
