+++ client/src/app/components/layout/main/home.component.scss
@@ -1,0 +1,11 @@
+.topAppBar {
+    z-index: 7;
+}
+
+.main-content {
+    width: 90%;
+}
+
+.drawer {
+    position: fixed;
+}
