+++ client/src/app/components/view/auth/auth.component.html
@@ -1,0 +1,14 @@
+<video autoplay muted loop class="background">
+  <source src="assets/login-background.mp4" type="video/mp4">
+</video>
+
+<div class="background" fxLayout="column" fxLayoutAlign="center">
+  <div fxLayout="row" fxLayoutAlign="center">
+      <div fxLayout="column" fxLayoutAlign="center" class="container" [mdcElevation]=4>
+        <div mdcHeadline4 class="login-header" >Welcome</div>
+        <mdc-text-field class="text-field" label="Username" required (change)="username=$event"></mdc-text-field>
+        <mdc-text-field class="text-field" label="Password" required (change)="password=$event"></mdc-text-field>
+        <button mdc-button primary raised (click)="onLogin()">Login</button>
+      </div>
+  </div>
+</div>
