+++ client/src/app/components/view/auth/auth.component.scss
@@ -1,0 +1,25 @@
+.container {
+    border-radius: 12px;
+    border: 1px solid rgba(0,0,0,.15);
+    padding: 30px;
+    margin-top: 50px;
+    background-color: rgba(256, 256, 256, .3);
+}
+
+.background {
+    position: fixed;
+    right: 0;
+    bottom: 0;
+    min-width: 100%;
+    min-height: 100%;
+}
+
+.login-header {
+    text-align: center;
+    padding-bottom: 20px;
+}
+
+.text-field {
+    background-color: rgba(256, 256, 256, .5);
+}
+
