+++ client/src/app/material.module.ts
@@ -11,6 +11,10 @@ import {
     MdcDialogModule,
     MdcTextFieldModule,
     MdcIconButtonModule,
+    MdcCardModule,
+    MdcTypographyModule,
+    MdcButtonModule,
+    MdcElevationModule,
  } from '@angular-mdc/web';
 
 @NgModule({
@@ -26,6 +30,10 @@ import {
     MdcFabModule,
     MdcDialogModule,
     MdcTextFieldModule,
+    MdcCardModule,
+    MdcTypographyModule,
+    MdcButtonModule,
+    MdcElevationModule,
   ]
 })
 export class MaterialModule { }
