+++ client/src/environments/env-defaults.ts
@@ -1,0 +1,3 @@
+export const EnvDefaults = {
+    jwtKeyName: "access_token",
+}
