+++ client/src/environments/environment.prod.ts
@@ -1,4 +1,7 @@
+import { EnvDefaults } from './env-defaults';
+
 export const environment = {
+  ...EnvDefaults,
   production: true,
   showErrors: false,
 };
