+++ client/src/environments/environment.ts
@@ -1,8 +1,11 @@
+import { EnvDefaults } from './env-defaults';
+
 // This file can be replaced during build by using the `fileReplacements` array.
 // `ng build --prod` replaces `environment.ts` with `environment.prod.ts`.
 // The list of file replacements can be found in `angular.json`.
 
 export const environment = {
+  ...EnvDefaults,
   production: false,
   showErrors: true,
 };
