+++ server/package.json
@@ -18,19 +18,25 @@
   "dependencies": {
     "@nestjs/common": "^6.0.0",
     "@nestjs/core": "^6.0.0",
+    "@nestjs/jwt": "^6.1.1",
     "@nestjs/microservices": "^6.6.7",
+    "@nestjs/passport": "^6.1.0",
     "@nestjs/platform-express": "^6.0.0",
     "@nestjs/swagger": "^3.1.0",
     "@nestjs/websockets": "^6.6.7",
     "dotenv": "^8.1.0",
     "morgan": "^1.9.1",
     "ncp": "^2.0.0",
+    "passport": "^0.4.0",
+    "passport-jwt": "^4.0.0",
+    "passport-local": "^1.0.0",
     "reflect-metadata": "^0.1.12",
     "rimraf": "^2.6.2",
     "rxjs": "^6.3.3",
     "swagger-ui-express": "^4.0.7"
   },
   "devDependencies": {
+    "@nestjs/cli": "^6.11.0",
     "@nestjs/testing": "^6.0.0",
     "@types/dotenv": "^6.1.1",
     "@types/express": "4.16.1",
@@ -38,6 +44,8 @@
     "@types/morgan": "^1.7.37",
     "@types/multer": "^1.3.10",
     "@types/node": "11.13.4",
+    "@types/passport-jwt": "^3.0.2",
+    "@types/passport-local": "^1.0.33",
     "@types/supertest": "2.0.7",
     "copy-webpack-plugin": "^5.0.4",
     "jest": "24.7.1",
