+++ server/src/routes.ts
@@ -3,4 +3,5 @@ export const joinRoutes = (...routes: string[]) => routes.join('/');
 export const routes: any = {
     api: "api",
     files: "files",
+    auth: "auth",
 }
