+++ server/src/services/config.service.ts
@@ -2,11 +2,13 @@ import { Injectable, Inject } from "@nestjs/common";
 import * as dotenv from 'dotenv';
 import * as path from 'path';
 import { ConfigMissingException } from '../lib/exceptions';
+import { randomBytes } from 'crypto';
 
 @Injectable()
 export class ConfigService {
     private readonly rootWildcard = '{.}';
     private readonly envFilePath: string;
+    private readonly _env: any;
     constructor(envFilePath: string) {
         this.envFilePath = envFilePath;
         dotenv.config({ path: path.join(this.envFilePath, 'default.env') }); 
@@ -16,10 +18,16 @@ export class ConfigService {
                 process.env[e] = val.replace(this.rootWildcard, this.envFilePath);
             }
         }
+        this._env = this.throwOnUndefined(process.env);
+        this.set('JWT_SECRET', randomBytes(20).toString())
     } 
 
     get env(): any {
-        return this.throwOnUndefined(process.env)
+        return this._env;
+    }
+
+    public async set(key: string, value: string) {
+        this._env[key] = value;
     }
     
     private throwOnUndefined(obj: any): string {
